requirements.txt
python-telegram-bot
openai
gspread
google-auth
python-dotenv
colorama
